% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center a tbl or data.frame at \code{xc}, \code{yc} \code{zc}}
\usage{
center(tbl, xc = NULL, yc = NULL, zc = NULL)
}
\arguments{
\item{tbl}{[data.frame or tibble], contains named columns \code{x}, \code{y}, \code{z}}

\item{xc}{[numeric], default NULL, length 1, new \code{x} center}

\item{yc}{[numeric], default NULL, length 1, new \code{y} center}

\item{zc}{[numeric], default NULL, length 1, new \code{z} center}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` centered at `xc`,`yc`,`zc`
}
}
\description{
Centers a table containing \code{x}, \code{y}, \code{z} coordinates at a point specified by
\code{xc}, \code{yc} \code{zc}. Using column \code{x} as an example, the new coordinates are
computed using \cr
\cr
\deqn{x_{i} = x_{i} - mean(x) + xc}
\cr
where \cr
\itemize{
\item \eqn{x}         is \verb{m x 1} column
\item \eqn{x_{i}}     is \code{i}th element in column
\item \eqn{mean(x)}   is \eqn{\frac{1}{\vert x \vert} \sum_i x_i}
\item \eqn{xc}        is \verb{1 x 1} scalar
}
}
