% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center3.R
\name{center3}
\alias{center3}
\title{Center pts in 3D}
\usage{
center3(
  tbl,
  at_x_ = NULL, at_y_ = NULL, at_z_ = NULL,
  x_ = x, y_ = y, z_ = z,
  use_bounding_box = TRUE,
  keep_bounding = TRUE
)
}
\arguments{
\item{tbl}{[tibble], a tibble containing points to be centered}

\item{at_x_}{[numeric], length 1, default NULL, \cr
value by which to center \code{x} variable in \code{tbl} \cr
default behavior is to leave center unchanged}

\item{at_y_}{[numeric], length 1, default NULL, \cr
value by which to center \code{y} variable in \code{tbl} \cr
default behavior is to leave center unchanged}

\item{at_z_}{[numeric], length 1, default NULL, \cr
value by which to center \code{z} variable in \code{tbl} \cr
default behavior is to leave center unchanged}

\item{x_}{[expr], variable name in \code{tbl} of \code{x} variable}

\item{y_}{[expr], variable name in \code{tbl} of \code{y} variable}

\item{z_}{[expr], variable name in \code{tbl} of \code{z} variable}

\item{use_bounding_box}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} rotate3 does not use bounding box points to make computations \cr
default behavior is to use bounding box points to make computations}

\item{keep_bounding}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} does not rbind the bounding box back after computation \cr
default behavior is to use bounding box points to make computations}
}
\value{
\itemize{
\item \link{tibble} with values
\item \code{x} centered x coordinate
\item \code{y} centered y coordinate
\item \code{z} centered z coordinate
\item \code{face} is point part of "front", "back", "left", "right", "bottom" or "top" face
\item \code{tb} is point on "top", "middle", "bottom" side of face
\item \code{rl} is point on "right", "middle", "left" side of face
\item \code{bounding_box} if \code{TRUE}, point is part of bounding box
}
}
\description{
{description placeholder}
}
