% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboid_slice.R
\name{cuboid_slice}
\alias{cuboid_slice}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
