% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypocycloid.R
\name{geom_hypocycloid}
\alias{geom_hypocycloid}
\alias{stat_hypocycloid}
\title{Draw a rose or hypocycloid curve}
\usage{
stat_hypocycloid(mapping = NULL, data = NULL, geom = "hypocycloid",
  position = "identity", n_points = 360, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

geom_hypocycloid(mapping = NULL, data = NULL, stat = "hypocycloid",
  position = "identity", n_points = 360, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{n_points}{The number of points to sample along the curve.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}
}
\description{
This geom allows you to draw the hypocycloid curve.
}
\details{
The curve follows the the parameterized form

\deqn{x = sin(a\theta + \delta)}
\deqn{y = sin(b\theta)}

these curves are closed when the radion \eqn{a / b} is rational. delta have
been scaled to be in the interval [0, 1].
}
\section{Aesthetics}{

geom_arc understand the following aesthetics (required aesthetics are in
bold):

- **x0**
- **y0**
- **a**
- **b**
- color
- fill
- size
- linetype
- alpha
- lineend
}

\section{Computed variables}{


\describe{
 \item{x, y}{The coordinates for the points along the rose curve}
}
}

\examples{
ggplot() +
  geom_hypocycloid(aes(a = 4, b = 1))

ggplot() +
  geom_hypocycloid(aes(a = 8, b = 1))

 ggplot() +
  geom_hypocycloid(aes(a = 2, b = 3, x0 = c(1, 4, 7)))
}
\references{
\url{https://en.wikipedia.org/wiki/Hypocycloid}
}
\author{
Emil Hvitfeldt
}
