% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_matrix_axes.R
\name{geom_matrix_axes}
\alias{geom_matrix_axes}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_viewport(aes(xc = 0, yc = 0, w = 2, h = 2)) +
  geom_matrix(
    mapping = aes(xc = 0, yc = 0, lx = 1, ly = 2, lz = 3),
    fill = 'skyblue3', colour = 'grey3',
    viewx = 2, viewy = 2
  ) +
  geom_matrix_axes(
    mapping = aes(xc = 0, yc = 0, lx = 1, ly = 2, lz = 3),
    arrow = arrow(15,unit(.25,'cm')),
    viewx = 2, viewy = 2
  )
}
