% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_matrix_face.R
\name{geom_matrix_face}
\alias{geom_matrix_face}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_matrix_face(
  mapping = aes(xc = 0, yc = 0, lx = 1, ly = 2, lz = 3),
  fill = 'skyblue3', colour = 'grey3',
  degx = 70, degy = 20, degz = 0,
  viewx = 2, viewy = 2, respect = TRUE
)
}
