% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_matrix_slice.R, R/geom_sparse_matrix.R
\name{geom_matrix_slice}
\alias{geom_matrix_slice}
\title{{title placeholder}}
\usage{
{usage placeholder}

{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}

\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}

{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_matrix_slice(
  mapping = aes(
    xc = 1, yc = 0,
    lx = 3, ly = 3, lz = 3,
    ix = 1, ixend = 1,
    iy = 1, iyend = 2,
    iz = 1, izend = 3),
  fill = 'skyblue3', colour = 'grey3',
  degx = 70, degy = 20, degz = 0,
  viewx = 2, viewy = 2, respect = TRUE
)
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_matrix_slice(
  mapping = aes(
    xc = 1, yc = 0,
    lx = 3, ly = 3, lz = 3,
    ix = 1, ixend = 1,
    iy = 1, iyend = 2,
    iz = 1, izend = 3),
  fill = 'skyblue3', colour = 'grey3',
  degx = 70, degy = 20, degz = 0,
  viewx = 2, viewy = 2, respect = TRUE
)
}
