% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_planar_circle.R
\name{geom_planar_circle}
\alias{geom_planar_circle}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


# unit circle
ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_planar_circle(mapping = aes(xc = 0, yc = 0, zc = 0, r = 1))
# equivalent to the default parameter call
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_planar_circle(mapping = aes(xc = 0, yc = 0, zc = 0, r = 1),degx = 0, degy = 0, degz = 0)
}
