% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_planar_ellipse.R
\name{geom_planar_ellipse}
\alias{geom_planar_ellipse}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
