% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_planar_hexagon.R
\name{geom_planar_hexagon}
\alias{geom_planar_hexagon}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_planar_hexagon(mapping = aes(xc = 0, yc = 0, zc = 0, r = 1))
}
