% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_planar_pointgon.R
\name{geom_planar_pointgon}
\alias{geom_planar_pointgon}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_planar_pointgon(
  data = pts_planar_circle(),
  mapping = aes(
    x = x, y = y, z = z
    )
  ) +
  geom_rect(
    mapping = aes(
      xmin = -1, xmax = +1,
      ymin = -1, ymax = +1
      ),
    fill = NA,
    colour = 'black'
  )
}
