% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{geom_rose}
\alias{geom_rose}
\alias{stat_rose}
\title{Draw a rose or rhodonea curve}
\usage{
stat_rose(mapping = NULL, data = NULL, geom = "rose",
  position = "identity", n_points = 360, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

geom_rose(mapping = NULL, data = NULL, stat = "rose",
  position = "identity", n_points = 360, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{n_points}{The number of points to sample along the curve.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}
}
\description{
This geom allows you to draw mathematical roses. A rose is a sinusoid
plotted in polar coordinates.
}
\details{
the shape follows the following parameterized form

\deqn{x = (cos(k\theta) + c)cos(\theta)}
\deqn{y = (cos(k\theta) + c)sin(\theta)}

where

\deqn{k = n / d}

the rose will be a closed loop when k is rational. this can easily be
achived by keeping n and d whole numbers. c is the internal offset parameter
changing how much the flower spreads out.
}
\section{Aesthetics}{

geom_arc understand the following aesthetics (required aesthetics are in
bold):

- **n**
- **d**
- **c**
- x0
- y0
- xscale
- yscale
- rotation
- color
- fill
- size
- linetype
- alpha
- lineend
}

\section{Computed variables}{


\describe{
 \item{x, y}{The coordinates for the points along the rose curve}
}
}

\examples{

ggplot() +
  geom_rose(aes(n = 2, d = 1, c = 0))

ggplot() +
  geom_rose(aes(n = 2, d = 8, c = 0))

ggplot() +
  geom_rose(aes(n = 5, d = 4, c = 4))

# Rescaling
ggplot() +
  geom_rose(aes(n = 5, d = 4, c = 4, xscale = 6, yscale = 2))

# Rotation
ggplot() +
  geom_rose(aes(n = 2, d = 1, c = 0, rotation = pi / 4))

# Multiple roses
ggplot() +
  geom_rose(aes(n = 5:1, d = 1:5, c = 0, x0 = 1:5 * 3))

ggplot() +
  geom_rose(aes(n = 5, d = 4, c = 4), fill = "white")
}
\references{
\url{https://en.wikipedia.org/wiki/Rose_(mathematics)}
}
\author{
Emil Hvitfeldt
}
