% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_unit_matrix.R
\name{geom_unit_matrix}
\alias{geom_unit_matrix}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_unit_matrix(mapping = aes(xc = 0, yc = 0),degx = 70, degy = 20, degz = 0)

# this matrix does not remain a matrix when aspect ratio of parent frame changes
ggplot() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_unit_matrix(mapping = aes(xc = .5, yc = 0),degx = 70, degy = 20, degz = 0)

# multiple rotation angles!
ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_unit_matrix(
  mapping = aes(xc = c(0,.5), yc = c(0,.5)),
  degx = c(70,20), degy = 20, degz = 0,
  fill = 'skyblue3'
  )
}
