% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_viewport.R
\name{geom_viewport}
\alias{geom_viewport}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
geom_viewport(
  data = tibble(xc = c(0,1),yc = c(0,1),w = c(.5,1), h = c(.5,1)),
  mapping = aes(xc = xc, yc = yc, w = w, h = h)
)
}
