% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomPlanarPoint.R
\docType{data}
\name{GeomPlanarPoint}
\alias{GeomPlanarPoint}
\title{{title placeholder}}
\format{
An object of class \code{GeomPlanarPoint} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case

}
\keyword{datasets}
