% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_cube_matrix.R
\name{pts_cube_matrix}
\alias{pts_cube_matrix}
\title{Unit matrix with labeled points}
\usage{
pts_cube_matrix(l)
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z`, `face`, `tb`, `rl`, `order`
}
}
\description{
Unit matrix with labeled points
}
\details{
\itemize{
\item \code{x}     is x coordinate of point
\item \code{y}     is y coordinate of point
\item \code{z}     is z coordinate of point
\item \code{face}  is face of point
\item \code{tb}    is tb is whether point is bottom or top, this is relative
\item \code{rl}    is rl is whether point is right or left, this is relative
\item \code{order} is for grouping in conjuction with \code{face} to pass to polygon
}
}
