% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_cuboid.R
\name{pts_cuboid}
\alias{pts_cuboid}
\title{Points of faces of cuboid}
\usage{
pts_cuboid(
  lx_ = 1,
  ly_ = 1,
  lz_ = 1
)
}
\arguments{
\item{lx_}{[numeric], length 1, non-negative, default 1, length of cuboid along x axis}

\item{ly_}{[numeric], length 1, non-negative, default 1, length of cuboid along y axis}

\item{lz_}{[numeric], length 1, non-negative, default 1, length of cuboid along z axis}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Points of faces of cuboid
}
