% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_cuboid_axis_label.R
\name{pts_cuboid_axis_label}
\alias{pts_cuboid_axis_label}
\title{Pts for cuboid axis labels}
\usage{
pts_cuboid_axis_label(
  lx_ = 1,
  ly_ = 1,
  lz_ = 1,
  expand_ = 1.2,
  face_ = 'top',
  side_ = 'left',
  label.start = "start",
  label.middle = "middle",
  label.end = "end",
  reverse = FALSE
)
}
\arguments{
\item{lx_}{[numeric], length 1, length of cuboid in x direction}

\item{ly_}{[numeric], length 1, length of cuboid in y direction}

\item{lz_}{[numeric], length 1, length of cuboid in z direction}

\item{expand_}{[numeric], length 1, nonnegative, proportion to expand axis by}

\item{face_}{[character], length 1, one of 'top','bottom','left','right','front', or 'back'}

\item{side_}{[character], length 1, one of 'left','right','top', or 'bottom'}

\item{label.start}{[character], length 1, default is "start", label for start of axis}

\item{label.middle}{[character], length 1, default is "middle", label for middle of axis}

\item{label.end}{[character], length 1, default is "end", label for end of axis}

\item{reverse}{[logical], length 1, default FALSE, if true axis is from top to bottom}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Pts for cuboid axis labels
}
