% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_cuboid_face.R
\name{pts_cuboid_face}
\alias{pts_cuboid_face}
\title{Points of a single face of a cuboid}
\usage{
pts_cuboid_face(
lx_ = 1,
ly_ = 1,
lz_ = 1,
face_ = 'bottom'
)
}
\arguments{
\item{lx_}{[numeric], length 1, non-negative, default 1, length of cuboid along x axis}

\item{ly_}{[numeric], length 1, non-negative, default 1, length of cuboid along y axis}

\item{lz_}{[numeric], length 1, non-negative, default 1, length of cuboid along z axis}

\item{face_}{[character], one of "top","bottom","left","right","front","back"}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Points of a single face of a cuboid
}
