% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_cuboid_slice.R
\name{pts_cuboid_slice}
\alias{pts_cuboid_slice}
\title{Points for a cuboid slice}
\usage{
pts_cuboid_slice(
  lx_ = 1, ly_ = 1, lz_ = 1,
  x_ = 0, xend_ = 1,
  z_ = 0, yend_ = 1,
  y_ = 0, zend_ = 1
)
}
\arguments{
\item{lx_}{[numeric], length 1, length of cuboid in \code{x} direction}

\item{ly_}{[numeric], length 1, length of cuboid in \code{y} direction}

\item{lz_}{[numeric], length 1, length of cuboid in \code{z} direction}

\item{x_}{[numeric], length 1, >= 0 & <= lx_ & <= xend_, start of slice in \code{x} direction}

\item{xend_}{[numeric], length 1, >= x_ & <= lx_, end of slice in \code{x} direction}

\item{z_}{[numeric], length 1, >= 0 & <= lz_ & <= zend_, start of slice in \code{z} direction}

\item{yend_}{[numeric], length 1, >= y_ & <= ly_, end of slice in \code{y} direction}

\item{y_}{[numeric], length 1, >= 0 & <= ly_ & <= yend_, start of slice in \code{y} direction}

\item{zend_}{[numeric], length 1, >= z_ & <= lz_, end of slice in \code{z} direction}
}
\value{
\itemize{
\item \link{tibble} with values
\item \code{x} coordinate of slice or bounding box
\item \code{y} coordinate of slice or bounding box
\item \code{z} coordinate of slice or bounding box
\item \code{face} is point part of "front", "back", "left", "right", "bottom" or "top" face
\item \code{tb} is point on "top", "middle", "bottom" side of face
\item \code{rl} is point on "right", "middle", "left" side of face
\item \code{bounding_box} if \code{TRUE}, point is part of bounding box
}
}
\description{
{description placeholder}
}
