% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_matrix_axis_label.R
\name{pts_matrix_axis_label}
\alias{pts_matrix_axis_label}
\title{Pts for matrix axis labels}
\usage{
pts_matrix_axis_label(
  nx_ = 1,
  ny_ = 1,
  nz_ = 1,
  expand_ = 1.2,
  face_ = "top",
  side_ = "left",
  label.start = "start",
  label.middle = "middle",
  label.end = "end",
  reverse = FALSE
)
}
\arguments{
\item{nx_}{[numeric], length 1, length of cells in x direction}

\item{ny_}{[numeric], length 1, length of cells in y direction}

\item{nz_}{[numeric], length 1, length of cells in z direction}

\item{expand_}{[numeric], length 1, nonnegative, proportion to expand axis by}

\item{face_}{[character], length 1, one of 'top','bottom','left','right','front', or 'back'}

\item{side_}{[character], length 1, one of 'left','right','top', or 'bottom'}

\item{label.start}{[character], length 1, default is "start", label for start of axis}

\item{label.middle}{[character], length 1, default is "middle", label for middle of axis}

\item{label.end}{[character], length 1, default is "end", label for end of axis}

\item{reverse}{[logical], length 1, default FALSE, if true axis is from top to bottom}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
#' @usage
pts_matrix_axis_label(
nx_ = 1,
ny_ = 1,
nz_ = 1,
expand_ = 1.2,
face_ = 'top',
side_ = 'left',
label.start = "start",
label.middle = "middle",
label.end = "end",
reverse = FALSE
)
}
