% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_arc.R
\name{pts_planar_arc}
\alias{pts_planar_arc}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_point(
    data = pts_planar_arc(),
    mapping = aes(x = x, y = y),
    size = 0.5
  ) +
  geom_polygon(
    data = pts_planar_arc(),
    mapping = aes(x = x, y = y),
    fill = 'NA',
    colour = 'black'
  )
# an arc
library(ggplot2)
library(tibble)
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_point(
    data = pts_planar_arc(a0 = 0, a1 = 90, n = 64 / 4),
    mapping = aes(x = x, y = y),
    size = 0.5
  ) +
  geom_path(
    data = pts_planar_arc(a0 = 0, a1 = 90, n = 64 / 4),
    mapping = aes(x = x, y = y),
    colour = 'black',
    arrow = arrow(angle = 15, length = unit(0.25,'cm'))
  )
# reversing the direction is as simple as reversing the start and end angles
library(ggplot2)
library(tibble)
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_point(
    data = pts_planar_arc(a0 = 90, a1 = 0, n = 64 / 4),
    mapping = aes(x = x, y = y),
    size = 0.5
  ) +
  geom_path(
    data = pts_planar_arc(a0 = 90, a1 = 0, n = 64 / 4),
    mapping = aes(x = x, y = y),
    colour = 'black',
    arrow = arrow(angle = 15, length = unit(0.25,'cm'))
  )
}
