% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_circle.R
\name{pts_planar_circle}
\alias{pts_planar_circle}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


# default is unit circle, z = 0
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_point(
    data = pts_planar_circle(n=32),
    mapping = aes(x = x, y = y)
  ) +
  geom_polygon(
    data = pts_planar_circle(n = 32),
    mapping = aes(x = x, y = y),
    fill = 'NA',
    colour = 'black'
  )

}
