% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_ngon.R
\name{pts_planar_ngon}
\alias{pts_planar_ngon}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


# default is unit circle, z = 0
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_polygon(
    data = pts_planar_ngon(n = 6),
    mapping = aes(x = x - 1, y = y - 1),
    fill = 'NA',
    colour = 'black'
  ) +
  geom_polygon(
    data = pts_planar_ngon(n = 4),
    mapping = aes(x = x, y = y),
    fill = 'NA',
    colour = 'black'
  ) +
  geom_polygon(
    data = pts_planar_ngon(n = 3),
    mapping = aes(x = x + 1, y = y + 1),
    fill = 'NA',
    colour = 'black'
  )
}
