% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_points.R
\name{pts_planar_points}
\alias{pts_planar_points}
\title{Pts for a planar points}
\usage{
pts_planar_points(
  pts1_ = c(0,0,1,1),
  pts2_ = c(0,1,0,1),
  c1_ = mean(pts1_),
  c2_ = mean(pts2_),
  l1_ = 1,
  l2_ = 1,
  p_ = 0,
  var1_ = x,
  var2_ = y,
  varp_ = z
)
}
\arguments{
\item{pts1_}{[numeric], length any, pts values axis 1}

\item{pts2_}{[numeric], length any, pts values on axis 2}

\item{c1_}{[numeric], length 1, center of points along axis 1, default mean(pts_l1_)}

\item{c2_}{[numeric], length 1, center of points along axis 2, default mean(pts_l2_)}

\item{l1_}{[numeric], length 1, length along axis 1, default is 1}

\item{l2_}{[numeric], length 1, length along axis 2, default is 1}

\item{p_}{[numeric], length 1, value for planar axis, default is 0}

\item{var1_}{[symbol], name of axis 1, default is x}

\item{var2_}{[symbol], name of axis 2, default is y}

\item{varp_}{[symbol], name of axis p, default is z}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Pts for a planar points
}
