% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_polygon.R
\name{pts_planar_polygon}
\alias{pts_planar_polygon}
\title{Create pts for a planar polygon}
\usage{
pts_planar_polygon(
  x = c(0,0,1,1),
  y = c(0,1,1,0),
  z = 0
)
}
\arguments{
\item{x}{[numeric], length >= 1, x coordinates of polygon}

\item{y}{[numeric], length >= 1, y coordinates of polygon}

\item{z}{[numeric], length >= 1, z coordinates of polygon}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Create pts for a planar polygon
}
