% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_scalar.R
\name{pts_scalar}
\alias{pts_scalar}
\title{Points for a Scalar Matrix}
\usage{
pts_scalar(
  keep_bounding = TRUE,
  faces_ = c('front','top','right')
)
}
\arguments{
\item{keep_bounding}{[logical], length 1 , default \code{TRUE}, \cr
if \code{FALSE} does not rbind the bounding box back after computation \cr
default behavior is to use bounding box points to make computations}

\item{faces_}{[character], one or more of 'front','back','top','bottom','left','right'}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Points for a Scalar Matrix
}
