% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_unit_bounding_box.R
\name{pts_unit_bounding_box}
\alias{pts_unit_bounding_box}
\title{Points for a unit bounding box}
\usage{
pts_unit_bounding_box()
}
\value{
\itemize{
\item \link{tibble} with values
\item \code{x} coordinate of slice or bounding box
\item \code{y} coordinate of slice or bounding box
\item \code{z} coordinate of slice or bounding box
\item \code{face} is point part of "front", "back", "left", "right", "bottom" or "top" face
\item \code{tb} is point on "top", "middle", "bottom" side of face
\item \code{rl} is point on "right", "middle", "left" side of face
\item \code{bounding_box} if \code{TRUE}, point is part of bounding box
}
}
\description{
Points for a unit bounding box
}
