% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_unit_viewport.R
\name{pts_unit_viewport}
\alias{pts_unit_viewport}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
