% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_viewport.R
\name{pts_viewport}
\alias{pts_viewport}
\title{Points for viewport}
\usage{
pts_viewport(
  xc_ = 0,
  yc_ = 0,
  s_ = 1
)
}
\arguments{
\item{xc_}{[numeric], length 1, non-negative, viewport x center, default is 0}

\item{yc_}{[numeric], length 1, non-negative, viewport x center, default is 0}

\item{s_}{[numeric], length 1, non-negative, amount to scale points by}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
Points for viewport
}
