% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotates x,y,z coords by specified degrees}
\usage{
rotate(tbl, degx = 0, degy = 0, degz = 0)
}
\arguments{
\item{degx}{[numeric], length 1, degrees to rotate \code{x} by}

\item{degy}{[numeric], length 1, degrees to rotate \code{y} by}

\item{degz}{[numeric], length 1, degrees to rotate \code{x} by}

\item{table}{[data.frame or tibble], contains named columns \code{x}, \code{y}, \code{z}}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` rotated by `degx`, `degy`, `degz`
}
}
\description{
Rotates \code{x}, \code{y}, \code{z} coordinates by degrees \code{degx}, \code{degy}, \code{degz} about the
origin (x = 0,y = 0,z = 0).
}
