% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rotate_df}
\alias{rotate_df}
\title{Rotates a parametric data.frame}
\usage{
rotate_df(df, theta)
}
\arguments{
\item{df}{a data frame with two columns x and y}

\item{theta}{a numeric denoting the amount of rotation}
}
\value{
a data frame with two columns x and y
}
\description{
Rotates a parametric data.frame
}
