% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale3.R
\name{scale3}
\alias{scale3}
\title{scales x,y,z coords by a specified amount}
\usage{
scale3(tbl, sx = 1, sy = 1, sz = 1, x = x, y = y, z = z, respect = FALSE)
}
\arguments{
\item{sx}{[numeric], length 1, nonnegative, amount to scale \code{x} by}

\item{sy}{[numeric], length 1, nonnegative, amount to scale \code{y} by}

\item{sz}{[numeric], length 1, nonngeative amount to scale \code{z} by}

\item{respect}{[logical], length 1, if true, scaling respect original aspect ratio}

\item{table}{[data.frame or tibble], contains named columns \code{x}, \code{y}, \code{z}}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` scale by `sx`, `sy`, `sz`
}
}
\description{
Scales \code{x}, \code{y}, \code{z} coordinates by amounts \code{sx}, \code{sy}, \code{sz} respectively.
Using column \code{x} as an example, the new coordinates are computed using \cr
\cr
\deqn{x_{i} = x_{i} * sx}
\cr
where \cr
\itemize{
\item \eqn{x}         is \verb{m x 1} column
\item \eqn{x_{i}}     is \code{i}th element in column
\item \eqn{sx}        is \verb{1 x 1} scalar
}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
#TODO
}
