% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_into.R
\name{scale_into}
\alias{scale_into}
\title{Scale a vector of values into a range}
\usage{
scale_into(x,
           new_xmin,
           new_xmax,
           old_xmin = NULL,
           old_xmax = NULL)
}
\arguments{
\item{x}{[numeric], length >= 1,}

\item{new_xmin}{[numeric], length == 1,}

\item{new_xmax}{[numeric], length == 1,}

\item{old_xmin}{[numeric], NULL or length 1}

\item{old_xmax}{[numeric], NULL or length 1}
}
\value{
\preformatted{      [numeric]
}
}
\description{
Scale a vector of values into a range
}
