% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_into_view3.R
\name{scale_into_view3}
\alias{scale_into_view3}
\title{scales x,y coords by to fit into a quasiviewport}
\usage{
scale_into_view3(
  tbl,
  mwx = NA,
  mwy = NA,
  w_var = x,
  h_var = y,
  old_w = NULL,
  old_h = NULL,
  x = x,
  y = y,
  respect = TRUE,
  use_bounding = FALSE
)
}
\arguments{
\item{tbl}{[data.frame or tibble], contains named columns \code{x}, \code{y}}

\item{mwx}{[numeric], length 1, nonnegative, maximum width of \code{x}}

\item{mwy}{[numeric], length 1, nonnegative, maximum width of \code{y}}

\item{respect}{[logical], length 1, if true (default), scaling respects original aspect ratio}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` scale by `sx`, `sy`, `sz`
}
}
\description{
#TODO
}
\note{

}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
#TODO

}
