% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_into_viewport.R
\name{scale_into_viewport}
\alias{scale_into_viewport}
\title{Scale a pts_tbl into a viewport}
\usage{
scale_into_viewport(
  tbl,
  xc_ = 0, yc_ = 0,
  side_ = 1,
  x_ = x, y_ = y,
  use_bounding_box = TRUE,
  keep_bounding = FALSE
)
}
\arguments{
\item{tbl}{[tibble], tibble of points}

\item{xc_}{[numeric], length 1, x center of viewport}

\item{yc_}{[numeric], length 1, y center of viewport}

\item{side_}{[numeric], length 1, nonnegative, the width of the side of the viewport}

\item{x_}{[symbol], length 1, variable name to use as \code{x} in \code{tbl}}

\item{y_}{[symbol], length 1, variable name to use as \code{y} in \code{tbl}}

\item{use_bounding_box}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} rotate3 does not use bounding box points to make computations \cr
default behavior is to use bounding box points to make computations}

\item{keep_bounding}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} does not rbind the bounding box back after computation \cr
default behavior is to use bounding box points to make computations}
}
\value{
\preformatted{      [\{type\}]
}
\itemize{
\item \link{tibble} with values
\item \code{x} x coordinate of viewport
\item \code{y} y coordinate of viewport
\item \code{z} z coordinate of viewport
\item \code{face} is point part of "front", "back", "left", "right", "bottom" or "top" face
\item \code{tb} is point on "top", "middle", "bottom" side of face
\item \code{rl} is point on "right", "middle", "left" side of face
}
}
\description{
Scale a pts_tbl into a viewport
}
