% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cuboid.R
\name{stat_cuboid}
\alias{stat_cuboid}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
stat_cuboid(mapping = aes(xc = 0, yc = 0, lx = 2, ly = 1, lz = 0.5))
}
