% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cuboid_slice.R
\name{stat_cuboid_slice}
\alias{stat_cuboid_slice}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
stat_cuboid_slice(
  mapping = aes(
    xc = 0, yc = 0,
    lx = 1, ly = 1, lz = 1,
    sx = .5, sxend = 1,
    sy = .2, syend = 1,
    sz = 0, szend = 1
    ),
  fill = NA,
  colour = 'black'
  )
}
