% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_matrix_slice_axes.R
\name{stat_matrix_slice_axes}
\alias{stat_matrix_slice_axes}
\title{{description placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
stat_matrix_slice_axes(mapping = aes(
  xc = 0, yc = 0,
  lx = 1, ly = 2, lz = 3,
  ix = 1, ixend = 1,
  iy = 1, iyend = 2,
  iz = 1, izend = 3),degx = 70, degy = 20, degz = 0)
}
