% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_planar_circle.R
\name{stat_planar_circle}
\alias{stat_planar_circle}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case



ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  stat_planar_circle(mapping = aes(xc = 0, yc = 0, zc = 0, r = 1))
}
