% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_planar_ngon.R
\name{stat_planar_ngon}
\alias{stat_planar_ngon}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case
}
