% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sparse_matrix.R
\name{stat_sparse_matrix}
\alias{stat_sparse_matrix}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# rotate a cube
#    70 degrees anticlockwise about its x axis
#    20 degrees anticlockwise about its y axis
# and then view a cube's front, right, and top faces
# notice that stat does not color the cube
# this is what geom_unit_cube is for!
ggplot() +
coord_equal() +
scale_x_continuous(limits = c(-2,2)) +
scale_y_continuous(limits = c(-2,2)) +
stat_matrix_slice(mapping = aes(
  xc = 1, yc = 0,
  lx = 3, ly = 3, lz = 3,
  ix = 1, ixend = 1,
  iy = 1, iyend = 2,
  iz = 1, izend = 3)
  ,degx = 70, degy = 20, degz = 0)
}
