% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_unit_matrix.R
\name{stat_unit_matrix}
\alias{stat_unit_matrix}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# rotate a cube
#    70 degrees anticlockwise about its x axis
#    20 degrees anticlockwise about its y axis
# and then view a cube's front, right, and top faces
# notice that stat does not color the cube
# this is what geom_unit_cube is for!
}
