% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_viewport.R
\name{stat_viewport}
\alias{stat_viewport}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


}
