% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMatrixSlice.R, R/StatSparseMatrix.R
\docType{data}
\name{StatMatrixSlice}
\alias{StatMatrixSlice}
\title{{title placeholder}}
\format{
An object of class \code{StatMatrixSlice} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatMatrixSlice} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
{usage placeholder}

{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}

\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}

{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


# rotate a matrix
#    70 degrees anticlockwise about its x axis
#    20 degrees anticlockwise about its y axis
# and then view a matrix's front, right, and top faces
}
\keyword{datasets}
