% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMatrixSliceLabels.R
\docType{data}
\name{StatMatrixSliceLabels}
\alias{StatMatrixSliceLabels}
\title{{title placeholder}}
\format{
An object of class \code{StatMatrixSliceLabels} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


}
\keyword{datasets}
