% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatUnitMatrix.R
\docType{data}
\name{StatUnitMatrix}
\alias{StatUnitMatrix}
\title{{title placeholder}}
\format{
An object of class \code{StatUnitMatrix} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\keyword{datasets}
