% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{translates x,y,z coords by a specified amount}
\usage{
translate(tbl, dx = 0, dy = 0, dz = 0)
}
\arguments{
\item{dx}{[numeric], length 1, amount to shift \code{x} by}

\item{dy}{[numeric], length 1, amount to shift \code{y} by}

\item{dz}{[numeric], length 1, amount to shift \code{z} by}

\item{table}{[data.frame or tibble], contains named columns \code{x}, \code{y}, \code{z}}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` shifted by `dx`, `dy`, `dz`
}
}
\description{
Moves \code{x}, \code{y}, \code{z} coordinates by amounts in \code{x}, \code{y}, \code{z} direction by
\code{dx}, \code{dy} \code{dz}. Using column \code{x} as an example, the new coordinates are
computed using \cr
\cr
\deqn{x_{i} = x_{i} + dx}
\cr
where \cr
\itemize{
\item \eqn{x}         is \verb{m x 1} column
\item \eqn{x_{i}}     is \code{i}th element in column
\item \eqn{dx}        is \verb{1 x 1} scalar
}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
#TODO


}
