% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{weave}
\alias{weave}
\title{Weaves two vectors together using alternating indecies}
\usage{
weave(a, b)
}
\arguments{
\item{a}{Vector}

\item{b}{Vector}
}
\value{
A vector
}
\description{
Weaves two vectors together using alternating indecies
}
