% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center3.R
\name{center3}
\alias{center3}
\title{Use expressions to center a tbl or data.frame at \code{xc}, \code{yc} \code{zc}}
\usage{
center3(tbl, xc = NULL, yc = NULL, zc = NULL, old_xc = NULL,
old_yc = NULL, old_zc = NULL, x = x, y = y, z = z)
}
\arguments{
\item{tbl}{[data.frame or tibble], contains named columns \code{x}, \code{y}, \code{z}}

\item{xc}{[numeric], default NULL, length 1, new \code{x} center}

\item{yc}{[numeric], default NULL, length 1, new \code{y} center}

\item{zc}{[numeric], default NULL, length 1, new \code{z} center}

\item{old_xc}{[numeric], default NULL, length 1, old \code{x} center}

\item{old_yc}{[numeric], default NULL, length 1, old \code{y} center}

\item{old_zc}{[numeric], default NULL, length 1, old \code{z} center}
}
\value{
\preformatted{      [tibble], with `x`, `y`, `z` centered at `xc`,`yc`,`zc`
}
}
\description{
Centers a table containing with variables aliased by \code{x}, \code{y}, \code{z}
at a point specified by expressions \code{xc}, \code{yc} \code{zc}. Using column
\code{x} as an example, the new coordinates are computed using \cr
\cr
\deqn{x_{i} = x_{i} - mean(x) + xc}
\cr
where \cr
\itemize{
\item \eqn{x}         is \verb{m x 1} column
\item \eqn{x_{i}}     is \code{i}th element in column
\item \eqn{mean(x)}   is \eqn{\frac{1}{\vert x \vert} \sum_i x_i}
\item \eqn{xc}        is \verb{1 x 1} scalar
}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# default call behavior is returns tibble or data.frame without
# modification (identity transformation)
tribble(
  ~x, ~y, ~z,
  0,  0,  0,
  0,  1,  0,
  1,  1,  0,
  1,  0,  0,
) \%>\%
  center3()

# fails if one or more of x,y,z not present
# modification (identity transformation)
# centering at 0, 0, 0
tribble(
  ~x, ~y, ~z,
  0,  0,  0,
  0,  1,  0,
  1,  1,  0,
  1,  0,  0,
) \%>\%
  center3(0,0,0)

# centering at 1, 0, 1
tribble(
  ~x, ~y, ~z,
  0,  0,  0,
  0,  1,  0,
  1,  1,  0,
  1,  0,  0,
) \%>\%
  center3(1,0,1)

# centering at 1, 0, 1
tribble(
  ~x, ~y, ~z,
  0,  0,  0,
  0,  1,  0,
  1,  1,  0,
  1,  0,  0,
) \%>\%
  center3(xc = 1,zc = 1)
}
