% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_matrix.R
\name{pts_matrix}
\alias{pts_matrix}
\title{Matrix points}
\usage{
pts_matrix(
  nx_ = 1,
  ny_ = 1,
  nz_ = 1
)
}
\arguments{
\item{nx_}{[integer], length 1, positive, x dimension of matrix}

\item{ny_}{[integer], length 1, positive, y dimension of matrix}

\item{nz_}{[integer], length 1, positive, z dimension of matrix}
}
\value{
\link{tibble} with columns
\itemize{
\item \code{x} x coordinate of point
\item \code{y} y coordinate of point
\item \code{z} z coordinate of point
\item \code{face} is face front, back, top, bottom, left or right
\item \code{tb} is the point on top, bottom or middle of face
\item \code{rl} is the point on left, right or middle of face
\item \code{point_order} points draw order
\item \code{face_order}  face draw order for default faces
}
}
\description{
{description placeholder}
}
