% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_planar_ellipse.R
\name{pts_planar_ellipse}
\alias{pts_planar_ellipse}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_point(
    data = pts_planar_ellipse(n=32),
    mapping = aes(x = x, y = y)
  ) +
  geom_polygon(
    data = pts_planar_ellipse(n = 32),
    mapping = aes(x = x, y = y),
    fill = 'NA',
    colour = 'black'
  )
ggplot() +
  coord_equal() +
  scale_x_continuous(limits = c(-2,2)) +
  scale_y_continuous(limits = c(-2,2)) +
  geom_polygon(
    data = pts_planar_ellipse(a = 1, b = 2, n=32),
    mapping = aes(x = x, y = y),
    fill = NA,
    colour = 'skyblue3'
  ) +
  geom_polygon(
    data = pts_planar_ellipse(a = 2, b = 1, n=32),
    mapping = aes(x = x, y = y),
    fill = NA,
    colour = 'orange2'
  )
}
