% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_unit_cube.R
\name{pts_unit_cube}
\alias{pts_unit_cube}
\title{{title placeholder}}
\usage{
{usage placeholder}
}
\arguments{
\item{{param}}{[{type}], {restrictions}}
}
\value{
\preformatted{      [\{type\}]
}
}
\description{
{description placeholder}
}
\examples{
library(dplyr)
library(magrittr)
library(ggplot2)
library(grid)
library(magrittr)
# identity case


}
