% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale3.R
\name{scale3}
\alias{scale3}
\title{Scale pts in 3D}
\usage{
scale3(
  tbl,
  sx_ = 1, sy_ = 1, sz_ = 1,
  x_ = x, y_ = y, z_ = z,
  use_bounding_box = TRUE,
  keep_bounding = TRUE
)
}
\arguments{
\item{tbl}{[tibble], a tibble containing points to be scaled}

\item{sx_}{[numeric], length 1, default 1, value by which to scale \code{x} variable in \code{tbl},}

\item{sy_}{[numeric], length 1, default 1, value by which to scale \code{y} variable in \code{tbl}}

\item{sz_}{[numeric], length 1, default 1, value by which to scale \code{z} variable in \code{tbl}}

\item{x_}{[expr], variable name in \code{tbl} of \code{x} variable}

\item{y_}{[expr], variable name in \code{tbl} of \code{y} variable}

\item{z_}{[expr], variable name in \code{tbl} of \code{z} variable}

\item{use_bounding_box}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} rotate3 does not use bounding box points to make computations \cr
default behavior is to use bounding box points to make computations}

\item{keep_bounding}{[logical], length 1, default \code{TRUE}, \cr
if \code{FALSE} does not rbind the bounding box back after computation \cr
default behavior is to use bounding box points to make computations}
}
\value{
\itemize{
\item \link{tibble} with values
\item \code{x} scaled x coordinate
\item \code{y} scaled y coordinate
\item \code{z} scaled z coordinate
\item \code{face} is point part of "front", "back", "left", "right", "bottom" or "top" face
\item \code{tb} is point on "top", "middle", "bottom" side of face
\item \code{rl} is point on "right", "middle", "left" side of face
\item \code{bounding_box} if \code{TRUE}, point is part of bounding box
}
}
\description{
Scale pts in 3D
}
